When tring to re-parse possible DNS responses, handle more exceptions
As it is, there are use-cases that involve taking blatantly non-DNS data
and parsing it with this code here. When that fails the first time, the
code chops off 2 bytes and tries again.  It is not surprising that this
eventually results in some exceptions due to unidentified parsing bugs
deep in the recursive guts of Net::DNS.

The issue is that not all of these exceptions are caught, so some bad
data may result in exceptions being thrown and in other cases not.  This
handles the `SystemStackError` I encountered several times while
testing and is in-line with how exceptions like this are handled
elsewhere in dap